requirements.txt
flask==1.1.1
wtforms==2.2.1
sendgrid==3.2.5
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn=20.0.4
pandas==1.0.4
pillow==7.0.0
